refactor(tools/explorer) Renaming lib and tests to path_explorer.
