+++ test/collection_template_sanity.py
@@ -2,7 +2,7 @@
 
 import unittest
 
-from test.test_tools.test_explorer import TestProjectExplorerLibrary
+from test.test_tools.test_path_explorer import TestProjectExplorerLibrary
 from test.test_tools.test_pretty_console import TestPrettyConsole
 
 if __name__ == '__main__':
