+++ test/test_tools/test_path_explorer.py
@@ -1,7 +1,7 @@
 from os.path import dirname
 from unittest import TestCase
 
-from tools import explorer
+from tools import path_explorer
 
 PATH_TO_HERE = dirname( __file__ )
 PATH_TO_TEST = dirname( PATH_TO_HERE )
@@ -13,7 +13,7 @@ class TestProjectExplorerLibrary( TestCase ):
         pass
 
     def test_explorer_on_test_directory( self ):
-        dirs, files = explorer.search_path_for_directories_and_files(
+        dirs, files = path_explorer.search_path_for_directories_and_files(
             path = PATH_TO_HERE,
             omit_python_cache = True
         )
@@ -29,7 +29,7 @@ class TestProjectExplorerLibrary( TestCase ):
         )
 
     def test_explorer_on_test_directory_including_cache( self ):
-        dirs, files = explorer.search_path_for_directories_and_files(
+        dirs, files = path_explorer.search_path_for_directories_and_files(
             path = PATH_TO_HERE,
             omit_python_cache = False
         )
@@ -40,45 +40,45 @@ class TestProjectExplorerLibrary( TestCase ):
         self.assertGreater( len( cache_folders_recorded ), 0 )
 
     def test_filtering_files_by_extension( self ):
-        files = explorer.search_path_for_files_with_extensions(
+        files = path_explorer.search_path_for_files_with_extensions(
             path = PATH_TO_TOOLS,
             extensions = ('py', 'xml')
         )
         self.assertGreaterEqual( len( files ), 3 )
 
-        files = explorer.search_path_for_files_with_extensions(
+        files = path_explorer.search_path_for_files_with_extensions(
             path = PATH_TO_HERE,
             extensions = ('py')
         )
         self.assertTrue( __file__ in files )
 
-        files = explorer.search_path_for_files_with_extensions(
+        files = path_explorer.search_path_for_files_with_extensions(
             path = PATH_TO_TOOLS,
             extensions = ('xml')
         )
         self.assertEqual( len( files ), 1 )
 
-        files = explorer.search_path_for_files_with_extensions(
+        files = path_explorer.search_path_for_files_with_extensions(
             path = PATH_TO_HERE,
             extensions = ()
         )
         self.assertEqual( len( files ), 0 )
 
-        files = explorer.search_path_for_files_with_extensions(
+        files = path_explorer.search_path_for_files_with_extensions(
             path = PATH_TO_HERE,
             extensions = ('com')
         )
         self.assertEqual( len( files ), 0 )
 
     def test_filtering_directories_by_partial_match( self ):
-        dirs = explorer.search_path_for_directories_with_partial_match(
+        dirs = path_explorer.search_path_for_directories_with_partial_match(
             path = dirname( PATH_TO_TEST ),
             partial = 'test'
         )
         self.assertGreaterEqual( len( dirs ), 1 )
         self.assertTrue( dirname( __file__ ) in dirs )
 
-        dirs = explorer.search_path_for_directories_with_partial_match(
+        dirs = path_explorer.search_path_for_directories_with_partial_match(
             path = dirname( PATH_TO_HERE ),
             partial = 'match me if you can'
         )
