+++ tools/path_explorer.py
